+++ src/components/functional/unlock-panel/components/UnlockPanelGroup/unlockPanelGroup.styles.ts
@@ -1,0 +1,9 @@
+// prettier-ignore
+export default {
+    unlockPanelGroupContainer: 'unlock-panel-group-container mvx:overflow-hidden mvx:flex mvx:sm:overflow-visible',
+    unlockPanelGroup: 'unlock-panel-group mvx:flex mvx:flex-col mvx:flex-1 mvx:overflow-hidden',
+    unlockPanelGroupLabel: 'unlock-panel-group-label mvx:leading-none mvx:relative mvx:text-secondary-text',
+    unlockPanelGroupProviders: 'unlock-panel-group-providers  mvx:flex mvx:flex-col mvx:gap-0.5 mvx:rounded-tl-lg mvx:rounded-tr-lg mvx:overflow-auto mvx:pb-6 mvx:mt-3 mvx:scrollbar-hide',
+    detectedPanelGroup: 'mvx:hidden mvx:sm:flex',
+    lastProviderButton: 'mvx:rounded-bl-lg mvx:rounded-br-lg mvx:overflow-hidden',
+} satisfies Record<string, string>;