+++ src/components/functional/unlock-panel/unlock-panel.tsx
@@ -7,16 +7,11 @@ import { ConnectionMonitor } from 'utils/ConnectionMonitor';
 import type { IEventBus } from 'utils/EventBus';
 import { EventBus } from 'utils/EventBus';
 
-import { UnlockPanelGroupSlotEnum } from './components/unlock-panel-group/unlock-panel-group';
 import { getIsExtensionAvailable, getIsMetaMaskAvailable } from './helpers';
 import type { IUnlockPanelManagerData } from './unlock-panel.types';
 import { UnlockPanelEventsEnum } from './unlock-panel.types';
-
-const unlockPanelClasses: Record<string, string> = {
-  detectedPanelGroup: 'mvx:hidden mvx:sm:flex',
-  desktopPanelGroupTitle: 'mvx:hidden mvx:sm:flex',
-  mobilePanelGroupTitle: 'mvx:sm:hidden',
-};
+import { UnlockPanelFooter, UnlockPanelGroup } from './components';
+import styles from './unlock-panel.styles';
 
 @Component({
   tag: 'mvx-unlock-panel',
@@ -175,7 +170,7 @@ export class UnlockPanel {
         <div
           id="anchor"
           ref={(element: HTMLDivElement) => this.setAnchor(element)}
-          class={{ 'unlock-panel-anchor': this.isLoggingIn || this.isIntroScreenVisible }}
+          class={{ [styles.unlockPanelAnchor]: this.isLoggingIn || this.isIntroScreenVisible }}
         >
           {this.isIntroScreenVisible && (
             <ProviderIdleScreen
@@ -188,33 +183,33 @@ export class UnlockPanel {
         </div>
 
         {isProviderScreenVisible && (
-          <div class="unlock-panel">
-            <div class="unlock-panel-groups">
+          <div class={styles.unlockPanel}>
+            <div class={styles.unlockPanelGroups}>
               {hasDetectedProviders && (
-                <mvx-unlock-panel-group
+                <UnlockPanelGroup
                   providers={detectedProviders}
-                  onLogin={(event: CustomEvent) => this.handleLogin(event.detail)}
-                  class={unlockPanelClasses.detectedPanelGroup}
-                >
-                  <div slot={UnlockPanelGroupSlotEnum.groupLabel}>Detected</div>
-                </mvx-unlock-panel-group>
+                  onLogin={(provider) => this.handleLogin(provider)}
+                  class={styles.detectedPanelGroup}
+                  groupLabel={
+                    <div>Detected</div>
+                  }
+                />
               )}
 
-              <mvx-unlock-panel-group
+              <UnlockPanelGroup
                 providers={otherProviders}
-                onLogin={(event: CustomEvent) => this.handleLogin(event.detail)}
-              >
-                <div slot={UnlockPanelGroupSlotEnum.groupLabel}>
-                  <div class={unlockPanelClasses.mobilePanelGroupTitle}>Options</div>
-                  <div class={unlockPanelClasses.desktopPanelGroupTitle}>
-                    {hasDetectedProviders ? 'Other Options' : 'Options'}
+                onLogin={(provider) => this.handleLogin(provider)}
+                groupLabel={
+                  <div>
+                    <div class={styles.mobilePanelGroupTitle}>Options</div>
+                    <div class={styles.desktopPanelGroupTitle}>
+                      {hasDetectedProviders ? 'Other Options' : 'Options'}
+                    </div>
                   </div>
-                </div>
-
-                <slot />
-              </mvx-unlock-panel-group>
+                }
+              />
             </div>
-            <mvx-unlock-panel-footer walletAddress={this.walletAddress} />
+            <UnlockPanelFooter walletAddress={this.walletAddress} />
           </div>
         )}
       </mvx-side-panel>